requirements.txt
numpy
scikit-learn
pillow
tensorflow
pytest
pandas
streamlit